Call Cell Hashing And Generate Report

Runs the default processing pipeline

CallAndGenerateReport(
  rawCountData,
  reportFile,
  callFile,
  barcodeWhitelist = NULL,
  cellbarcodeWhitelist = "inputMatrix",
  methods = c("multiseq", "htodemux"),
  citeSeqCountDir = NULL,
  minCountPerCell = 5,
  title = NULL,
  metricsFile = NULL,
  skipNormalizationQc = FALSE
)

Arguments

rawCountData
The input barcode file or umi_count folder

reportFile
The file to which the HTML report will be written

callFile
The file to which the table of calls will be written

barcodeWhitelist
A vector of barcode names to retain.

cellbarcodeWhitelist
Either a vector of expected barcodes (such as all cells with passing gene expression data), or the string 'inputMatrix'. If the latter is provided, the set of cellbarcodes present in the original unfiltered count matrix will be stored and used for reporting. This allows the report to count cells that were filtered due to low counts separately from negative/non-callable cells.

methods
The set of methods to use for calling. See GenerateCellHashingCalls for options.

citeSeqCountDir
This is the root folder of the Cite-seq-Count output, containing umi_count and read_count folders. If provided, this will be used to generate a library saturation plot

minCountPerCell
Cells (columns) will be dropped if their total count is less than this value.

title
A title for the HTML report

metricsFile
If provided, summary metrics will be written to this file.

skipNormalizationQc
If true, the normalization/QC plots will be skipped. These can be time consuming on large input data.