Generate Cell Hashing Calls

The primary methods to generating cell hashing calls from a filtered matrix of count data.

GenerateCellHashingCalls(
  barcodeMatrix,
  methods = c("bff_cluster", "multiseq", "dropletutils"),
  methodsForConsensus = NULL,
  cellbarcodeWhitelist = NULL,
  metricsFile = NULL,
  doTSNE = TRUE,
  doHeatmap = TRUE,
  perCellSaturation = NULL,
  ...
)

Arguments

barcodeMatrix

The filtered matrix of hashing count data

methods

A vector of one or more calling methods to use. Currently supported are: htodemux, multiseq, dropletutils, gmm_demux, demuxem, bff_raw, and bff_cluster

methodsForConsensus

By default, a consensus call will be generated using all methods; however, if this parameter is provided, all algorithms specified by methods will be run, but only the list here will be used for the final consensus call. This allows one to see the results of a given caller without using it for the final calls.

cellbarcodeWhitelist

A vector of expected cell barcodes. This allows reporting on the total set of expected barcodes, not just those in the filtered count matrix.

metricsFile

If provided, summary metrics will be written to this file.

doTSNE

If true, tSNE will be run on the resulting hashing calls after each caller. This can be useful as a sanity check; however, adds time.

doHeatmap

If true, Seurat::HTOHeatmap will be run on the results of each caller. Not supported by all callers.

perCellSaturation

An optional dataframe with the columns cellbarcode and saturation. This will be merged into the final output.

...

Caller-specific arguments can be passed by prefixing with the method name. For example, htodemux.positive.quantile = 0.95, will be passed to the htodemux positive.quantile argument).

Value

A data frame of results.