Plot Library Saturation

Create a plot of the library saturation per cell

PlotLibrarySaturation(citeseqCountDir, metricsFile = NULL)

Arguments

citeseqCountDir

The root of the Cite-seq-Count output folder, which should contain umi_count and read_count folders.

metricsFile

If provided, summary metrics will be written to this file.

Value

The overall saturation for this library